Update README.md
The purpose of this project is to analyse the company’s end to end operations. The primary objective is to work closely with various teams like operations, support and marketing to derive valuable insights from various datasets to improve operational efficiency and understand sudden changes in key metrics and also using advanced SQL skills to investigate and explain metric spikes such as fluctuation in user engagement or sales figures. The analysis will help organization to proactively address the challenges and ultimately leading to continuous growth and improvement.